move prepare scripts into room_scaper
